Cleaning old structure
