+++ lua/lazy.lua
@@ -1,0 +1,8 @@
+local M = {}
+
+function M.setup(config)
+
+  
+end
+
+return M
