--- lua/narya/default_config.lua
@@ -1,14 +1,0 @@
-local config = {
-    alpha_enabled = true,
-    neo_tree_enabled = true,
-    bufferline_enabled = true,
-    disable_narya_keybindings = false,
-    narya_memo = {
-        "https://twitter.com/lc_fd",
-        "https://github.com/lcfd?tab=repositories",
-        "https://strawberry.rocks/docs",
-        "https://stackoverflow.com/"
-    }
-}
-
-return config
