--- lua/narya/lazy.lua
@@ -1,25 +1,0 @@
-local M = {}
-
-function M.setup(config)
-  -- Install Lazy.nvim
-  local lazypath = vim.fn.stdpath("data") .. "/lazy/lazy.nvim"
-
-  if not vim.loop.fs_stat(lazypath) then
-    vim.fn.system({
-      "git",
-      "clone",
-      "--filter=blob:none",
-      "https://github.com/folke/lazy.nvim.git",
-      "--branch=stable", -- latest stable release
-      lazypath,
-    })
-  end
-
-  vim.opt.rtp:prepend(lazypath)
-
-  local opts = {}
-
-  require("lazy").setup('plugins', opts)
-end
-
-return M
