+++ lua/plugins/git.lua
@@ -1,0 +1,13 @@
+return {
+  {
+    "TimUntersberger/neogit",
+    dependencies = "nvim-lua/plenary.nvim",
+    config = function(_, opts)
+      require("neogit").setup({
+        commit_popup = {
+          kind = "split",
+        },
+      })
+    end,
+  },
+}
