+++ lua/plugins/stimpack.lua
@@ -1,0 +1,52 @@
+return {
+  {
+    "ggandor/leap.nvim",
+    keys = {
+      { "s", mode = { "n", "x", "o" }, desc = "Leap forward to" },
+      { "S", mode = { "n", "x", "o" }, desc = "Leap backward to" },
+      { "gs", mode = { "n", "x", "o" }, desc = "Leap from windows" },
+    },
+    config = function(_, opts)
+      local leap = require("leap")
+      for k, v in pairs(opts) do
+        leap.opts[k] = v
+      end
+      leap.add_default_mappings(true)
+      vim.keymap.del({ "x", "o" }, "x")
+      vim.keymap.del({ "x", "o" }, "X")
+    end,
+  },
+  {
+    "numToStr/Comment.nvim",
+    config = function(_, opts)
+      require("Comment").setup({
+        toggler = {
+          ---Line-comment toggle keymap
+          line = "<leader>1",
+          ---Block-comment toggle keymap
+          block = "<leader>2",
+        },
+        opleader = {
+          ---Line-comment keymap
+          line = "<leader>1",
+          ---Block-comment keymap
+          block = "<leader>2",
+        },
+      })
+    end,
+  },
+  {
+    "kylechui/nvim-surround",
+    version = "*",
+    event = "VeryLazy",
+    config = function()
+      require("nvim-surround").setup({})
+    end,
+  },
+  {
+    "johmsalas/text-case.nvim",
+  },
+  {
+    "nanotee/zoxide.vim",
+  },
+}
