+++ lua/plugins/tests.lua
@@ -1,0 +1,23 @@
+return {
+  {
+    "nvim-neotest/neotest",
+    lazy = true,
+    dependencies = {
+      "nvim-lua/plenary.nvim",
+      "nvim-treesitter/nvim-treesitter",
+      "antoinemadec/FixCursorHold.nvim",
+    },
+    config = function()
+      require("neotest").setup({
+        adapters = {
+          require("neotest-python"),
+        },
+        runner = "pytest",
+        python = ".venv/bin/python",
+      })
+    end,
+  },
+  {
+    "nvim-neotest/neotest-python",
+  },
+}
